VCV-247: Refactored charts so the pie chart doesn't refresh when user toggles between chart types
